[v0.9.8.3-types5] rm serious warnings for changing player_t. Others remain so far.

+++ mytypes.h
@@ -76,7 +76,7 @@
 
 typedef int64_t gamesnum_t;
 
-typedef int32_t player_t;
+typedef int64_t player_t;
 
 struct gamei {
 	player_t 	whiteplayer;
